requirements.txt
fastapi
uvicorn
groq
pandas
openpyxl
sqlparse
python-dotenv
